Start splitting project into multiple files
